converted InventoryContext to typescript
